requirements.txt
colorama==0.4.6
dill==0.4.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.4
pygments==2.19.2
pyyaml==6.0.3
rich==14.2.0
scipy==1.17.1
taichi==1.7.4
